fixed deletion of dyno messages

+++ Create_Team.py
@@ -2,7 +2,7 @@ import discord
 import random
 import json
 import time
-
+import subprocess
 
 
 # functions
@@ -28,59 +28,6 @@ def create_team(players):
 
 def read_json():
     return json.load(open('configuration.json', 'r'))
-
-# def set_constants():
-#     constants_input = json.load(open('configuration.json', 'r'))
-#     #commands
-#     global COMMAND_CREATE_TEAM
-#     global COMMAND_PLAY_LOL
-#     
-#     #channel
-#     global CHANNEL_CREATE_TEAM_VOICE
-#     global CHANNEL_CREATE_TEAM_TEXT
-#     global CHANNEL_PLAY_REQUESTS
-#     
-#     #toggle
-#     global TOGGLE_AUTO_DELETE
-#     global TOGGLE_COMMAND_ONLY
-#     global TOGGLE_AUTO_REACT
-#     global TOGGLE_AUTO_DM
-# 
-#     #message
-#     global MESSAGE_TEAM_1
-#     global MESSAGE_TEAM_2
-# 
-#     #misc
-#     global AUTO_REACT_PATTERN
-#     global VERSION
-#     global AUTO_REACT_PASS_EMOJI
-#     global NOTIFY_ON_REACT_PURGE_TIMER
-#     
-#     # order: top, jgl, mid, adc, supp, fill
-#     global EMOJI_ID_LIST
-# 
-#     COMMAND_CREATE_TEAM = constants_input["COMMAND_CREATE_TEAM"]
-#     COMMAND_PLAY_LOL = constants_input["COMMAND_PLAY_LOL"]
-#     
-#     CHANNEL_CREATE_TEAM_VOICE =  constants_input["CHANNEL_CREATE_TEAM_VOICE"]
-#     CHANNEL_CREATE_TEAM_TEXT = constants_input["CHANNEL_CREATE_TEAM_TEXT"]
-#     CHANNEL_PLAY_REQUESTS= constants_input["CHANNEL_PLAY_REQUESTS"]
-# 
-#     TOGGLE_AUTO_DELETE = constants_input["TOGGLE_AUTO_DELETE"]
-#     TOGGLE_COMMAND_ONLY = constants_input["TOGGLE_COMMAND_ONLY"]
-#     TOGGLE_AUTO_REACT = constants_input["TOGGLE_AUTO_REACT"]
-#     TOGGLE_AUTO_DM = constants_input["TOGGLE_AUTO_DM"]
-# 
-#     MESSAGE_TEAM_1 = constants_input["MESSAGE_TEAM_1"]
-#     MESSAGE_TEAM_2 = constants_input["MESSAGE_TEAM_2"]
-# 
-#     AUTO_REACT_PATTERN = constants_input["AUTO_REACT_PATTERN"]
-#     VERSION = constants_input["VERSION"]
-
-
-
-
-
 
 # init
 client = discord.Client()
@@ -164,12 +111,10 @@ async def on_message(message):
             await message.channel.send("Done.")
 
      
-     # deletes messages that are not commands in channel create-play-requests
+     # deletes all messages that are not commands(except bot responses) in channel create-play-requests
     elif message.content.startswith(constants["COMMAND_PLAY_LOL"]) == False and (str(message.channel) == str(constants["CHANNEL_PLAY_REQUESTS"])):
-        await message.delete()
-    # elif message.content.startswith('!end'):
-    #     await message.channel.send('Bye!')
-    #     await  client.logout()
+        if(message.author.name != "Dyno"):
+            await message.delete()
         
    
 # automatically dms user if a reaction in NOTIFIY_ON_REACT_CHANNEL was added with a NOTIFY_ON_REACT_PURGE_TIMER delay
@@ -183,7 +128,6 @@ async def on_reaction_add(reaction, user):
     scheduled_purge_for_notifiy_on_react()
     message_sender_id = int((reaction.message.content.split(None, 1)[1]).split(None,1)[0][3:-1])
     message_sender = client.get_user(message_sender_id)
-    #or str(reaction.message.channel.name) == 'bot'
     if(constants["TOGGLE_AUTO_DM"] and str(reaction.message.channel.name) == constants["CHANNEL_PLAY_REQUESTS"] and message_sender != user and user not in user_delay_cache):
         if reaction.message.author.name == "Dyno":
             #only works for the specfied message format: '@everyone @user [rest of msg]'
